+++ packages/backend/src/core/NoteUpdateService.ts
@@ -9,7 +9,7 @@ import { Inject, Injectable, OnApplicationShutdown } from '@nestjs/common';
 import { extractCustomEmojisFromMfm } from '@/misc/extract-custom-emojis-from-mfm.js';
 import { extractHashtags } from '@/misc/extract-hashtags.js';
 import { MiNote, IMentionedRemoteUsers } from '@/models/Note.js';
-import type { NotesRepository, DriveFilesRepository, UsersRepository } from '@/models/_.js';
+import type { NotesRepository, DriveFilesRepository, UsersRepository, UserNotePiningsRepository } from '@/models/_.js';
 import type { MiDriveFile } from '@/models/DriveFile.js';
 import type { MiUser, MiLocalUser, MiRemoteUser } from '@/models/User.js';
 import { DI } from '@/di-symbols.js';
@@ -23,6 +23,7 @@ import { UserEntityService } from '@/core/entities/UserEntityService.js';
 import { RelayService } from '@/core/RelayService.js';
 import { In } from 'typeorm';
 import ActiveUsersChart from '@/core/chart/charts/active-users.js';
+import { FeaturedCollectionCacheService } from '@/core/FeaturedCollectionCacheService.js';
 import util from 'util';
 
 export type NoteUpdateData = {
@@ -49,6 +50,9 @@ export class NoteUpdateService implements OnApplicationShutdown {
 		@Inject(DI.driveFilesRepository)
 		private driveFilesRepository: DriveFilesRepository,
 
+		@Inject(DI.userNotePiningsRepository)
+		private userNotePiningsRepository: UserNotePiningsRepository,
+
 		private globalEventService: GlobalEventService,
 		private searchService: SearchService,
 		private apRendererService: ApRendererService,
@@ -56,6 +60,7 @@ export class NoteUpdateService implements OnApplicationShutdown {
 		private userEntityService: UserEntityService,
 		private relayService: RelayService,
 		private activeUsersChart: ActiveUsersChart,
+		private featuredCollectionCacheService: FeaturedCollectionCacheService,
 	) {
 	}
 
@@ -128,6 +133,13 @@ export class NoteUpdateService implements OnApplicationShutdown {
 		// Fetch the updated note
 		const updatedNote = await this.notesRepository.findOneByOrFail({ id: note.id });
 
+		if (this.userEntityService.isLocalUser(user) && await this.userNotePiningsRepository.existsBy({
+			userId: user.id,
+			noteId: note.id,
+		})) {
+			await this.featuredCollectionCacheService.invalidate(user.id);
+		}
+
 		if (!silent) {
 			if (this.userEntityService.isLocalUser(user)) this.activeUsersChart.write(user);
 
@@ -141,7 +153,6 @@ export class NoteUpdateService implements OnApplicationShutdown {
 			//#region AP deliver
 			if (this.userEntityService.isLocalUser(user)) {
 				setImmediate(async () => {
-					// @ts-ignore
 					const noteActivity = await this.renderNoteActivity(updatedNote, user);
 					await this.deliverToConcerned(user, updatedNote, noteActivity);
 				});